refacto heal objects (#3378)
* refacto heal objects

* refacto model preparation options

* test obj3d only

* merge master

* remove heal and simplify from m3d

* fix docstrings

* add more tests

* moved solve example to solvers run

* linux CI_CD

---------
