+++ _unittest/test_07_Object3D.py
@@ -630,3 +630,36 @@ class TestClass:
         self.aedtapp.modeler.create_box([60, 0, 0], [5, 10, 10], matname="vacuum")
         nets = self.aedtapp.identify_touching_conductors()
         assert len(nets) == 2
+
+    def test_62_heal_objects(self):
+        self.aedtapp.insert_design("Heal_Objects")
+        self.aedtapp.modeler.create_box([0, 1.5, 0], [1, 2.5, 5], name="box_1")
+        self.aedtapp.modeler.create_box([0, 1.5, 0], [1, 2.5, 5], name="box_2")
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1")
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1,box_2")
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1, box_2 ")
+        assert not self.aedtapp.modeler.heal_objects(input_objects_list=["box_1", "box_2"])
+        assert not self.aedtapp.modeler.heal_objects(input_objects_list="box_1", simplify_type=3)
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", max_stitch_tolerance="0.01")
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", max_stitch_tolerance=0.01)
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", geometry_simplification_tolerance=1.2)
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", geometry_simplification_tolerance="1.2")
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", tighten_gaps_width=0.001)
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", tighten_gaps_width="0.001")
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", silver_face_tolerance=1.2)
+        assert self.aedtapp.modeler.heal_objects(input_objects_list="box_1", silver_face_tolerance="1.2")
+        assert not self.aedtapp.modeler.heal_objects(input_objects_list=None)
+        assert not self.aedtapp.modeler.heal_objects(input_objects_list=1)
+
+    def test_20_simplify_objects(self):
+        assert self.aedtapp.modeler.simplify_objects(input_objects_list="box_1")
+        assert self.aedtapp.modeler.simplify_objects(input_objects_list="box_1,box_2")
+        assert self.aedtapp.modeler.simplify_objects(input_objects_list="box_1, box_2")
+        assert not self.aedtapp.modeler.simplify_objects(input_objects_list=["box_1", "box_2"])
+        assert self.aedtapp.modeler.simplify_objects(input_objects_list="box_1", simplify_type="Primitive Fit")
+        assert not self.aedtapp.modeler.simplify_objects(input_objects_list="box_1", simplify_type="Invalid")
+        assert not self.aedtapp.modeler.simplify_objects(
+            input_objects_list="box_1", simplify_type="Polygon Fit", extrusion_axis="U"
+        )
+        assert not self.aedtapp.modeler.simplify_objects(input_objects_list=None)
+        assert not self.aedtapp.modeler.simplify_objects(input_objects_list=1)
