+++ _unittest/test_27_Maxwell2D.py
@@ -406,47 +406,6 @@ class TestClass:
         assert "var_test" in self.aedtapp.variable_manager.design_variable_names
         assert self.aedtapp.variable_manager.design_variables["var_test"].expression == "234"
 
-    def test_29_heal_objects(self):
-        self.aedtapp.set_active_design("design_for_test")
-        assert self.aedtapp.heal_objects(input_objects_list="Rotor_Section1")
-        assert self.aedtapp.heal_objects(input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1")
-        assert self.aedtapp.heal_objects(input_objects_list="Rotor_Section1, Magnet1_Section1, Magnet2_Section1 ")
-        assert not self.aedtapp.heal_objects(input_objects_list="")
-        assert not self.aedtapp.heal_objects(
-            input_objects_list=["Rotor_Section1", "Magnet1_Section1", "Magnet2_Section1"]
-        )
-        assert not self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", simplify_type=3
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", max_stitch_tolerance="0.01"
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", max_stitch_tolerance=0.01
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", geometry_simplification_tolerance=1.2
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1",
-            geometry_simplification_tolerance="1.2",
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", tighten_gaps_width=0.001
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", tighten_gaps_width="0.001"
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", silver_face_tolerance=1.2
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="Rotor_Section1,Magnet1_Section1,Magnet2_Section1", silver_face_tolerance="1.2"
-        )
-
-    def test_30_simplify_objects(self):
-        assert not self.aedtapp.simplify_objects(input_objects_list="Rotor_Section1")
-
     def test_31_cylindrical_gap(self):
         assert not self.aedtapp.mesh.assign_cylindrical_gap("Band")
         [
