+++ _unittest/test_28_Maxwell3D.py
@@ -697,22 +697,6 @@ class TestClass:
         box3 = self.aedtapp.modeler.create_box([-50, -50, -50], [1, 1, 1], matname="copper")
         assert len(self.aedtapp.get_conduction_paths()) == 2
 
-    def test_42_harmonic_forces(self, m3dtransient):
-        assert m3dtransient.enable_harmonic_force(
-            ["Stator"],
-            force_type=2,
-            window_function="Rectangular",
-            use_number_of_last_cycles=True,
-            last_cycles_number=3,
-            calculate_force="Harmonic",
-        )
-        m3dtransient.save_project()
-        m3dtransient.analyze(m3dtransient.active_setup, num_cores=2)
-        assert m3dtransient.export_element_based_harmonic_force(
-            start_frequency=1, stop_frequency=100, number_of_frequency=None
-        )
-        assert m3dtransient.export_element_based_harmonic_force(number_of_frequency=5)
-
     def test_43_eddy_effect_transient(self, m3dtransient):
         assert m3dtransient.eddy_effects_on(["Rotor"], activate_eddy_effects=True)
 
@@ -794,39 +778,6 @@ class TestClass:
         self.aedtapp["var_test"] = "234"
         assert "var_test" in self.aedtapp.variable_manager.design_variable_names
         assert self.aedtapp.variable_manager.design_variables["var_test"].expression == "234"
-
-    def test_47_heal_objects(self):
-        self.aedtapp.set_active_design("Motion")
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box")
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box,impedance_box_copper,Inner_Box")
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box, impedance_box_copper, Inner_Box ")
-        assert not self.aedtapp.heal_objects(input_objects_list=["impedance_box", "impedance_box_copper", "Inner_Box"])
-        assert not self.aedtapp.heal_objects(input_objects_list="impedance_box", simplify_type=3)
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box_copper", max_stitch_tolerance="0.01")
-        assert self.aedtapp.heal_objects(input_objects_list="Inner_Box", max_stitch_tolerance=0.01)
-        assert self.aedtapp.heal_objects(
-            input_objects_list="impedance_box,Inner_Box", geometry_simplification_tolerance=1.2
-        )
-        assert self.aedtapp.heal_objects(
-            input_objects_list="impedance_box,Inner_Box", geometry_simplification_tolerance="1.2"
-        )
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box,Inner_Box", tighten_gaps_width=0.001)
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box,Inner_Box", tighten_gaps_width="0.001")
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box,Inner_Box", silver_face_tolerance=1.2)
-        assert self.aedtapp.heal_objects(input_objects_list="impedance_box,Inner_Box", silver_face_tolerance="1.2")
-
-    def test_48_simplify_objects(self):
-        assert self.aedtapp.simplify_objects(input_objects_list="impedance_box")
-        assert self.aedtapp.simplify_objects(input_objects_list="impedance_box,impedance_box_copper,Inner_Box")
-        assert self.aedtapp.simplify_objects(input_objects_list="impedance_box, impedance_box_copper, Inner_Box ")
-        assert not self.aedtapp.simplify_objects(
-            input_objects_list=["impedance_box", "impedance_box_copper", "Inner_Box"]
-        )
-        assert self.aedtapp.simplify_objects(input_objects_list="impedance_box", simplify_type="Primitive Fit")
-        assert not self.aedtapp.simplify_objects(input_objects_list="impedance_box", simplify_type="Invalid")
-        assert not self.aedtapp.simplify_objects(
-            input_objects_list="impedance_box", simplify_type="Polygon Fit", extrusion_axis="U"
-        )
 
     def test_49_cylindrical_gap(self, cyl_gap):
         [
